requirements.txt
asgiref==3.7.2
Django==4.1.13
djangorestframework==3.14.0
Pygments==2.14.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
